requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
CacheControl==0.12.6
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
contextlib2==21.6.0
distlib==0.3.3
distro==1.6.0
Django==3.2.7
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==21.1
git+https://github.com/jazzband/django-silk.git#egg=django-silk
django-watchman==1.1.1
djangorestframework==3.12.4
gprof2dot==2021.2.21
gunicorn==20.1.0
html5lib==1.1
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
ipaddr==2.2.0
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.2
packaging==21.0
pep517==0.11.0
pluggy==1.0.0
progress==1.6
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.1
requests==2.26.0
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.5.0
